+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/arbeidsplassen/form/initialValues.tsx
@@ -196,46 +196,3 @@ export const initialKursVerdier = {
 
 export const initialSammendragVerdi =
 	'Trivelig person med kjærlighet for sau, som har tonnevis med allsidig kompetanse og fantastiske personlige egenskaper.'
-
-export const initialDigitalInnsending = {
-	tittel: '',
-	tema: '',
-	kanal: 'NAV_NO',
-	sak: {
-		sakstype: 'GENERELL_SAK',
-		fagsaksystem: '',
-		fagsakId: '',
-	},
-	ferdigstill: true,
-	avsenderMottaker: {
-		id: '',
-		navn: '',
-		idType: '',
-	},
-	journalfoerendeEnhet: undefined,
-	dokumenter: [
-		{
-			tittel: '',
-			brevkode: '',
-		},
-	],
-}
-
-export const initialDokarkiv = {
-	tittel: runningE2ETest() ? 'test' : '',
-	tema: runningE2ETest() ? 'test ' : '',
-	sak: {
-		sakstype: 'GENERELL_SAK',
-		fagsaksystem: '',
-		fagsakId: '',
-	},
-	kanal: 'SKAN_IM',
-	ferdigstill: true,
-	journalfoerendeEnhet: undefined,
-	dokumenter: [
-		{
-			tittel: runningE2ETest() ? 'test' : '',
-			brevkode: runningE2ETest() ? 'test' : '',
-		},
-	],
-}
