+++ apps/dolly-frontend/src/main/js/src/components/fagsystem/dokarkiv/form/partials/Digitalinnsending.tsx
@@ -3,19 +3,19 @@ import { FormSelect } from '@/components/ui/form/inputs/select/Select'
 import { FormTextInput } from '@/components/ui/form/inputs/textInput/TextInput'
 import { Kategori } from '@/components/ui/form/kategori/Kategori'
 
-export default () => {
+export default ({ path }: { path: string }) => {
 	return (
 		// @ts-ignore
 		<div className="flexbox--space">
 			<Kategori title={'Avsender'}>
 				<div className="flexbox--flex-wrap">
 					<FormSelect
-						name={'dokarkiv.avsenderMottaker.idType'}
+						name={`${path}.avsenderMottaker.idType`}
 						label={'Type'}
 						options={Options('avsenderType')}
 					/>
-					<FormTextInput name="dokarkiv.avsenderMottaker.id" label="ID" size={'large'} />
-					<FormTextInput name="dokarkiv.avsenderMottaker.navn" label="Navn" />
+					<FormTextInput name={`${path}.avsenderMottaker.id`} label="ID" size={'large'} />
+					<FormTextInput name={`${path}.avsenderMottaker.navn`} label="Navn" />
 				</div>
 			</Kategori>
 		</div>
